pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "streambind"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.24.0",
    "av>=16.0.1",
    "bgutil-ytdlp-pot-provider>=1.2.2",
    "numpy>=2.4.0",
    "pillow>=12.0.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "yt-dlp>=2025.12.8",
]

[project.scripts]
streambind = "app.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "Q",
    "D100",
    "D101",
    "D107",
    "D103",
    "D102",
    "D104",
    "S311",
    "F403",
    "F405",
    "INP001",
    "G003",
    "RUF001",
    "RUF002",
    "RUF003",
    "EM101",
    "EM102",
    "F401",
    "PGH003",
    "ANN401",
    "TC002",
    "TRY003",
]
